+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/model/AssetBaseEntity.java
@@ -24,8 +24,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.eclipse.tractusx.traceability.assets.domain.model.Owner;
-import org.eclipse.tractusx.traceability.assets.domain.model.QualityType;
+import org.eclipse.tractusx.traceability.assets.domain.asbuilt.model.Owner;
+import org.eclipse.tractusx.traceability.assets.domain.asbuilt.model.QualityType;
 
 @NoArgsConstructor
 @Getter